Add Logging Lib
I added a logging library and a supporting color library. This logging
library adds basic logging functions like info, debug, and err. The
logging outputs a standard time stamp, the type of logging, and the
message. The different logging functions such as info err are in
distinct colors. Naturally err is red and info in blue.