Add Common
I added a file for common functions. To start with I added an example
for a find loop. I find my self doing this on ocasion and needing to
constantly look up how to do it. I would like to have it take a function
so that I can have some work done in the loop.