Repeted Code into Function
I was using a formatted date a lot so I created a function for it. At
first I put it on the common.sh since it was a common function. However,
this caused a circular depenency and broke things. I ended up putting
this function into the logging.sh file where it is used the most. I may
put copy this function and put it into the common.sh file later if
needed. I then replaced all accurances of the date format with the new
function. Making things much cleaner and readable.

I also added a very basic README.